Refactor REST routes into dedicated controllers
