+++ supersede-css-jlg-enhanced/src/Infra/Rest/ControllerInterface.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace SSC\Infra\Rest;
+
+interface ControllerInterface
+{
+    public function registerRoutes(): void;
+}
