+++ supersede-css-jlg-enhanced/src/Infra/Rest/SystemController.php
@@ -1,0 +1,42 @@
+<?php declare(strict_types=1);
+
+namespace SSC\Infra\Rest;
+
+use WP_REST_Response;
+
+final class SystemController extends BaseController
+{
+    public function registerRoutes(): void
+    {
+        register_rest_route('ssc/v1', '/health', [
+            'methods' => 'GET',
+            'permission_callback' => [$this, 'authorizeRequest'],
+            'callback' => [$this, 'healthCheck'],
+        ]);
+    }
+
+    public function healthCheck(): WP_REST_Response
+    {
+        $assets_to_check = [
+            'css/admin.css', 'css/ux.css', 'js/ux.js', 'js/utilities.js',
+            'codemirror/lib/codemirror.js', 'codemirror/lib/codemirror.css'
+        ];
+        $asset_status = [];
+        foreach ($assets_to_check as $asset) {
+            $asset_status[$asset] = is_file(SSC_PLUGIN_DIR . 'assets/' . $asset) ? 'OK' : 'Manquant';
+        }
+
+        $data = [
+            'plugin_version' => defined('SSC_VERSION') ? SSC_VERSION : 'N/A',
+            'wordpress_version' => get_bloginfo('version'),
+            'php_version' => phpversion(),
+            'rest_api_status' => 'OK',
+            'asset_files_exist' => $asset_status,
+            'composer_dependencies' => [
+                'Minify' => class_exists('\\MatthiasMullie\\Minify\\CSS') ? 'Chargé' : 'Non trouvé',
+            ]
+        ];
+
+        return new WP_REST_Response($data, 200);
+    }
+}
